+++ core/src/main/java/org/apache/mahout/cf/taste/recommender/RecommendedItem.java
@@ -32,6 +32,6 @@ public interface RecommendedItem extends Comparable<RecommendedItem> {
    *
    * @return strength of the preference
    */
-  double getValue();
+  float getValue();
 
 }
